Made it so that budget components lives inside budgets top component
